+++ src/Symfony/Component/HttpKernel/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 7.2
 ---
 
+ * Add call to `Request::resetFormats()` in `Kernel::boot()` to reset the request formats at the start of a new main request.
  * Remove `@internal` flag and add `@final` to `ServicesResetter`
  * Add support for `SYMFONY_DISABLE_RESOURCE_TRACKING` env var
  * Add support for configuring trusted proxies/headers/hosts via env vars